tbrekalo-build: update CMake and create source file for algorithms
